pyproject.toml
[project]
name = "kennapartner-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Emmanuel",email = "emmanuelhilary9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "kennapartner_backend", from = "src"}]

[tool.poetry.scripts]
seed = "kennapartner_backend.utils.seed:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

